requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.4
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
whitenoise==6.8.2
